bot-service/requirements.txt
aiogram==3.7.0
httpx==0.27.0
python-dotenv==1.0.1
aiohttp==3.9.5
